fix: layout duplo, persistência pagamento, simplificação ProductForm, QR Code Pix
- App.tsx: rotas /produtos/novo e /produtos/:id/editar usam P (sem DashboardLayout externo)
  pois ProductForm já inclui DashboardLayout internamente — elimina duplo wrapper

- drizzle/schema.ts + 0014_payment_settings_links.sql: adiciona campos de links globais
  (payment_platform, pix_key, pix_link, card_payment_url, boleto_url) à tabela
  permupay_payment_settings com ADD COLUMN IF NOT EXISTS (idempotente)

- drizzle/meta/_journal.json: registra migration 0014 no journal para que o
  scripts/migrate.mjs aplique automaticamente no próximo deploy

- server/routers.ts: remove .int() de cardInstallments e boletoMonths no
  paymentSettingsSchema (banco usa real, não integer)

- server/db.payment-settings.ts: DEFAULTS completos com novos campos de links

- client/src/pages/ProductForm.tsx: remove seção 'Links de Pagamento' e todos
  os campos (paymentPlatform, pixKey, pixLink, cardPaymentUrl, boletoUrl) do
  FormState, defaultForm, useEffect de defaults globais e handleSave payload —
  links agora são responsabilidade exclusiva da ConfiguracoesPagamento global

- client/src/pages/ConfiguracoesPagamento.tsx: adiciona QR Code Pix gerado
  automaticamente a partir da chave Pix configurada (qrcode.react v4.2.0)

- client/src/components/ImageUpload.tsx: corrige onSuccess(publicUrl) → onSuccess(finalUrl)
- client/src/pages/Marketplace.tsx: corrige [...new Set()] → Array.from(new Set())
- server/db.orders.ts: remove chamada a activateNextBatchIfNeeded (função inexistente)

+++ server/routers.ts
@@ -129,13 +129,13 @@ const paymentSettingsSchema = z.object({
   cardDebitFee: z.number().min(0).max(100).optional(),
   cardCreditCashFee: z.number().min(0).max(100).optional(),
   cardCreditInstallmentFee: z.number().min(0).max(100).optional(),
-  cardInstallments: z.number().int().min(1).max(48).optional(),
+  cardInstallments: z.number().min(1).max(48).optional(),
   cardAnticipationRate: z.number().min(0).max(100).optional(),
   cardMonthlyRate: z.number().min(0).max(100).optional(),
   cardCustomerPaysInterest: z.boolean().optional(),
 
   // Boleto
-  boletoMonths: z.number().int().min(1).max(60).optional(),
+  boletoMonths: z.number().min(1).max(60).optional(),
   boletoMonthlyRate: z.number().min(0).max(100).optional(),
   boletoFixedFee: z.number().min(0).optional(),
   boletoDefaultRisk: z.number().min(0).max(100).optional(),
